refactor: mover loginService para subpasta user e atualizar imports
